# LumenServiceApp/public/swaggerdoc.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lumen API",
        "license": {
            "name": "TEDC"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "summary": "Register a new user",
                "operationId": "86b18331ccd0a9be778980dafcf8a33d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "admin"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "User Name",
                                    "email": "user@gmail.com",
                                    "password": "123456",
                                    "password_confirmation": "123456",
                                    "role": "admin"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/public/post": {
            "get": {
                "summary": "Get Public posts",
                "operationId": "ef02a97f8f773d5bb3dcf6f826433f54",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "summary": "Create a new post",
                "operationId": "c818baff7bb6540e41fb8fd8e1f24ef5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "video": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "students_id": {
                                        "type": "string"
                                    },
                                    "categories_id": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "Post Title",
                                    "content": "Post Content",
                                    "image": "Post Image",
                                    "video": "Post Video",
                                    "status": "draft",
                                    "students_id": "1",
                                    "categories_id": "1",
                                    "user_id": "1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/public/post/{id}": {
            "get": {
                "summary": "Get Public post by id",
                "operationId": "c1774b27934606a1f63c096d4e5f4274",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}